lib: move List and Queue into library
