+++ Makefile.am
@@ -48,7 +48,7 @@ neomutt_SOURCES = account.c addrbook.c address.h alias.c alias.h attach.c \
 	curs_lib.c curs_main.c edit.c editmsg.c enter.c enter_state.h \
 	envelope.c envelope.h filter.c flags.c format_flags.h from.c getdomain.c group.c \
 	handler.c hdrline.c header.h headers.c help.c history.c hook.c \
-	init.c keymap.c list.h main.c mbox.c mbyte.c mbtable.h \
+	init.c keymap.c main.c mbox.c mbyte.c mbtable.h \
 	menu.c mh.c muttlib.c mutt_idna.c mutt_sasl_plain.c mutt_socket.c \
 	mutt_tunnel.c mx.c newsrc.c nntp.c options.h pager.c parameter.c parameter.h \
 	parse.c pattern.c pattern.h pop.c pop_auth.c pop_lib.c postpone.c \
@@ -88,7 +88,7 @@ EXTRA_DIST = account.h attach.h bcache.h browser.h buffy.h \
 	mutt_curses.h mutt_idna.h mutt_lua.h mutt_menu.h mutt_notmuch.h \
 	mutt_options.h mutt_regex.h mutt_sasl.h mutt_sasl_plain.h \
 	mutt_socket.h mutt_ssl.h mutt_tunnel.h mx.h myvar.h nntp.h opcodes.h pager.h \
-	pgpewrap.c pop.h protos.h queue.h README.md README.SSL remailer.c remailer.h \
+	pgpewrap.c pop.h protos.h README.md README.SSL remailer.c remailer.h \
 	rfc1524.h rfc2047.h rfc2231.h rfc3676.h rfc822.h sidebar.h \
 	sort.h txt2c.c txt2c.sh version.h mutt_tags.h
