+++ hcache/Makefile.am
@@ -44,12 +44,12 @@ BUILT_SOURCES = $(HCVERSION)
 $(top_srcdir)/keymap_defs.h:
 	+$(MAKE) -C $(top_srcdir) keymap_defs.h
 
-hcversion.h: $(top_srcdir)/mutt.h $(top_srcdir)/address.h $(top_srcdir)/list.h \
+hcversion.h: $(top_srcdir)/mutt.h $(top_srcdir)/address.h $(top_srcdir)/lib/list.h \
 		$(top_srcdir)/lib/buffer.h $(top_srcdir)/parameter.h \
 		$(top_srcdir)/body.h $(top_srcdir)/envelope.h \
 		$(top_srcdir)/header.h $(srcdir)/hcachever.sh
 	( echo '#include "config.h"'; echo '#include "mutt.h"'; \
-	echo '#include "address.h"'; echo '#include "list.h"'; \
+	echo '#include "address.h"'; echo '#include "lib/list.h"'; \
 	echo '#include "lib/buffer.h"'; echo '#include "parameter.h"'; \
 	echo '#include "body.h"'; echo '#include "envelope.h"'; \
 	echo '#include "header.h"'; \
