+++ hook.c
@@ -35,7 +35,6 @@
 #include "envelope.h"
 #include "globals.h"
 #include "header.h"
-#include "list.h"
 #include "mailbox.h"
 #include "mutt_regex.h"
 #include "ncrypt/ncrypt.h"
