+++ imap/util.c
@@ -45,7 +45,6 @@
 #include "globals.h"
 #include "header.h"
 #include "imap/imap.h"
-#include "list.h"
 #include "mailbox.h"
 #include "message.h"
 #include "mutt_socket.h"
