+++ mbox.c
@@ -42,7 +42,6 @@
 #include "envelope.h"
 #include "globals.h"
 #include "header.h"
-#include "list.h"
 #include "mailbox.h"
 #include "mutt_curses.h"
 #include "mx.h"
