+++ mutt_tags.h
@@ -25,7 +25,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
-#include "queue.h"
+#include "lib/queue.h"
 
 /**
  * struct TagNode - LinkedList Tag Element
