+++ muttlib.c
@@ -55,7 +55,6 @@
 #include "format_flags.h"
 #include "globals.h"
 #include "header.h"
-#include "list.h"
 #include "mailbox.h"
 #include "mime.h"
 #include "mutt_curses.h"
