+++ ncrypt/crypt_mod.c
@@ -23,7 +23,7 @@
 #include "config.h"
 #include "lib/lib.h"
 #include "crypt_mod.h"
-#include "queue.h"
+#include "lib/queue.h"
 
 /**
  * struct CryptModule - A crypto plugin module
