+++ postpone.c
@@ -38,7 +38,6 @@
 #include "globals.h"
 #include "header.h"
 #include "keymap.h"
-#include "list.h"
 #include "mailbox.h"
 #include "mime.h"
 #include "mutt_menu.h"
