+++ remailer.c
@@ -40,7 +40,6 @@
 #include "globals.h"
 #include "header.h"
 #include "keymap.h"
-#include "list.h"
 #include "mutt_curses.h"
 #include "mutt_menu.h"
 #include "opcodes.h"
